requirements.txt

pandas>=2.2.0
requests>=2.32.0
python-dotenv>=1.0.1
plotly>=5.24.0
